+++ apps/docs/src/components/demo/ThemeCustomizer.tsx
@@ -9,7 +9,7 @@ import { Label } from '../ui/Label'
 import { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from '../ui/Select'
 
 export function ThemeCustomizer({ className, ...props }: ComponentProps<'div'>) {
-  const { state, updateState } = useT3M4('demo')
+  const { state, setState } = useT3M4('demo')
 
   return (
     <div className={cn('flex w-full items-center gap-2', className)} {...props}>
@@ -20,9 +20,9 @@ export function ThemeCustomizer({ className, ...props }: ComponentProps<'div'>) 
             key={color}
             variant='link'
             size='sm'
-            data-active={color === state.base?.facets.color}
-            disabled={!state.base || !!state.forced?.facets?.color}
-            onClick={() => updateState({ facets: { color } })}
+            data-active={color === state.computed?.facets.color}
+            disabled={!state.computed || !!state.forced?.facets?.color}
+            onClick={() => setState({ facets: { color } })}
             className='text-muted-foreground hover:text-primary data-[active=true]:text-primary mr-4 cursor-pointer text-base capitalize transition-colors hover:no-underline'
           >
             {color}
@@ -36,7 +36,7 @@ export function ThemeCustomizer({ className, ...props }: ComponentProps<'div'>) 
         <Label htmlFor='theme-selector' className='sr-only'>
           Theme
         </Label>
-        <Select disabled={!state.base || !!state.forced?.facets?.color} value={state.base?.facets.color ?? config.demo.facets.color.default} onValueChange={(color: T3M4<'demo'>['facets']['color']) => updateState({ facets: { color } })}>
+        <Select disabled={!state.computed || !!state.forced?.facets?.color} value={state.computed?.facets.color ?? config.demo.facets.color.default} onValueChange={(color: T3M4<'demo'>['facets']['color']) => setState({ facets: { color } })}>
           <SelectTrigger id='theme-selector' size='sm' className='capitalize shadow-none *:data-[slot=select-value]:w-12'>
             <span className='font-medium'>Theme:</span>
             <SelectValue placeholder='Select a theme' />